[FidelityFX] rename AMDFidelityFX/ folder to AMD/
