+++ Shaders/AMD/DNSR/LICENSE.txt
