+++ Shaders/AMD/DNSR/ffx_denoiser_reflections_common.h
