+++ Shaders/AMD/DNSR/ffx_denoiser_reflections_config.h
