+++ Shaders/AMD/DNSR/ffx_denoiser_reflections_resolve_temporal.h
