+++ Shaders/AMD/DNSR/ffx_denoiser_shadows_filter.h
