+++ Shaders/AMD/DNSR/ffx_denoiser_shadows_prepare.h
