+++ Source/Renderer/CMakeLists.txt
@@ -10,22 +10,22 @@ set (LibHeadersD3D12X
     "Libs/D3DX12/d3dx12.h"
 )
 set (LibFFX_CACAO
-    "Libs/AMDFidelityFX/CACAO/ffx_cacao_defines.h"
-    "Libs/AMDFidelityFX/CACAO/ffx_cacao.cpp"
-    "Libs/AMDFidelityFX/CACAO/ffx_cacao.h"
-    "Libs/AMDFidelityFX/CACAO/ffx_cacao_impl.cpp"
-    "Libs/AMDFidelityFX/CACAO/ffx_cacao_impl.h"
+    "Libs/AMD/CACAO/ffx_cacao_defines.h"
+    "Libs/AMD/CACAO/ffx_cacao.cpp"
+    "Libs/AMD/CACAO/ffx_cacao.h"
+    "Libs/AMD/CACAO/ffx_cacao_impl.cpp"
+    "Libs/AMD/CACAO/ffx_cacao_impl.h"
 )
 set (LibFFX_SSSR
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_128spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_16spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_1spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_256spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_2spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_32spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_4spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_64spp.cpp"
-    "Libs/AMDFidelityFX/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_8spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_128spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_16spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_1spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_256spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_2spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_32spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_4spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_64spp.cpp"
+    "Libs/AMD/SSSR/samplerCPP/samplerBlueNoiseErrorDistribution_128x128_OptimizedFor_2d2d2d2d_8spp.cpp"
 )
 set (LibD3D12MA
     "Libs/D3D12MA/src/D3D12MemAlloc.h"
@@ -166,33 +166,33 @@ set (Shaders
     "../../Shaders/VQPlatform.h"
 )
 set (FFX_CAS_Shaders
-    "../../Shaders/AMDFidelityFX/CAS/ffx_a.h"
-    "../../Shaders/AMDFidelityFX/CAS/ffx_cas.h"
+    "../../Shaders/AMD/CAS/ffx_a.h"
+    "../../Shaders/AMD/CAS/ffx_cas.h"
 )
 set (FFX_FSR_Shaders
-    "../../Shaders/AMDFidelityFX/FSR1.0/ffx_a.h"
-    "../../Shaders/AMDFidelityFX/FSR1.0/ffx_fsr1.h"
+    "../../Shaders/AMD/FSR1.0/ffx_a.h"
+    "../../Shaders/AMD/FSR1.0/ffx_fsr1.h"
 )
 set (FFX_SPD_Shaders
-    "../../Shaders/AMDFidelityFX/SPD/ffx_a.h"
-    "../../Shaders/AMDFidelityFX/SPD/ffx_spd.h"
+    "../../Shaders/AMD/SPD/ffx_a.h"
+    "../../Shaders/AMD/SPD/ffx_spd.h"
 )
 set (FFX_CACAO_Shaders
-    "../../Shaders/AMDFidelityFX/CACAO/ffx_cacao.hlsl"
+    "../../Shaders/AMD/CACAO/ffx_cacao.hlsl"
 )
 set (FFX_SSSR_Shaders
-    "../../Shaders/AMDFidelityFX/SSSR/ffx_sssr.h"
+    "../../Shaders/AMD/SSSR/ffx_sssr.h"
 )
 set (FFX_DNSR_Shaders
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_reflections_common.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_reflections_config.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_reflections_prefilter.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_reflections_reproject.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_reflections_resolve_temporal.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_shadows_filter.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_shadows_prepare.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_shadows_tileclassification.h"
-    "../../Shaders/AMDFidelityFX/DNSR/ffx_denoiser_shadows_util.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_reflections_common.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_reflections_config.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_reflections_prefilter.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_reflections_reproject.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_reflections_resolve_temporal.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_shadows_filter.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_shadows_prepare.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_shadows_tileclassification.h"
+    "../../Shaders/AMD/DNSR/ffx_denoiser_shadows_util.h"
 )
 set (ShadersScreenSpaceReflections
     "../../Shaders/ScreenSpaceReflections/ClassifyReflectionTiles.hlsl"
@@ -278,4 +278,4 @@ target_include_directories(${PROJECT_NAME} PUBLIC "../../") # VQE root
 target_include_directories(${PROJECT_NAME} PUBLIC "./") # Renderer root
 
 add_custom_command(TARGET ${PROJECT_NAME} PRE_BUILD
-    COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/Libs/AMDFidelityFX/CACAO/build_shaders_dxil.bat)
+    COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/Libs/AMD/CACAO/build_shaders_dxil.bat)
