+++ Source/Renderer/Rendering/RenderPass/AmbientOcclusion.h
@@ -21,7 +21,7 @@
 
 // #define FFX_CACAO_ENABLE_PROFILING 1
 // #define FFX_CACAO_ENABLE_NATIVE_RESOLUTION 1 
-#include "Renderer/Libs/AMDFidelityFX/CACAO/ffx_cacao.h"
+#include "Renderer/Libs/AMD/CACAO/ffx_cacao.h"
 
 #include <DirectXMath.h>
 #include <dxgiformat.h>
